refactor: promote Actorconf, urlRemoteFromInforesid to broader namespace
