+++ tr_sys/tr_ars/default_ars_app/api.py
@@ -8,6 +8,7 @@ from django.http import HttpResponse
 from django.urls import reverse
 from django.views.decorators.csrf import csrf_exempt
 from tr_smartapi_client.smart_api_discover import SmartApiDiscover
+from utils2 import Actorconf, urlRemoteFromInforesid
 
 logger = logging.getLogger(__name__)
 
@@ -94,42 +95,8 @@ def init_api_fn(actorconf):
     fn.__doc__ = "Forward api request at %s to %s" % (fn.__name__, actorconf.inforesid())
     return fn
 
-class Actorconf:
-    def __init__(self, inforesid,  name, path, method, params) -> None:
-        self._inforesid = inforesid
-        self._name = name
-        self._path = path
-        self._method = method
-        self._params = params
-
-    def inforesid(self):
-        return self._inforesid
-
-    def name(self):
-        return self._name
-    
-    def path(self):
-        return self._path
-    
-    def method(self):
-        return self._method
-    
-    def params(self):
-        return self._params
-
-
 def make_actorconf(inforesid, name, path, method=None, params=None):
     return Actorconf(inforesid, name, path, method, params)
-
-def urlRemoteFromInforesid(inforesid):
-    urlServer=SmartApiDiscover().urlServer(inforesid)
-    if urlServer is not None:
-        endpoint=SmartApiDiscover().endpoint(inforesid)
-        params=SmartApiDiscover().params(inforesid)
-        return (urlServer +
-                (("/"+endpoint) if endpoint is not None else "") +
-                (("?"+params) if params is not None else "")) if urlServer is not None else None
-    return None
 
 def init_api_index(actors, app_path):
     def fn(req):
